v0.1.5
Created a credits scene which is currently the same as the how to play scene.

Music now transitions somewhat smoothly. Menu music will play through the main menu, loading, how to play, and credits scene. The menu music will stop if the game scene is loaded. The menu music will start again when moving from the game scene to the menu.