v0.1.13
Created the loading screen asset which randomly selects an enemy to display along with its text. Imported more assets. Worked on game over status, made it so enemies no longer move and towers stop shooting after the game has ended. The play can not interact with anything if the game is over.