v0.0.19
Quality of Life Improvements
Added a visualization to show which tower is currently selected.
Added tooltips to all modules in the inventory to show rarity and affect.

Bug Fixes
Fixed an issue with dragging modules while the final enemy of a round is destroyed causing the round to not end.
Fixed an issue with the placement tower attached to the mouse appearing behind UI elements.
Fixed more issues with enemies having incorrect statistics.